feat(employee): summary row, detail navigation, shared dialogs, form sheet
- EmployeeFormModal: converted from MUI Dialog to FormSheet (slide-in drawer)
- EmployeeDialogs: new shared container for form/payment/delete dialogs, driven by employeeStore.dialogMode
- EmployeePage: adds EmployeeSummary row, navigates to detail on row click via useNavigate, removes direct dialog imports in favour of EmployeeDialogs
