+++ src/pages/EmployeePage.tsx
@@ -1,107 +1,52 @@
-import React, { useEffect, useMemo } from "react";
-import EmployeeFormModal from "components/employee/Form/EmployeeFormModal";
+import React, { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
+import EmployeeDialogs from "components/employee/EmployeeDialogs";
+import EmployeeSummary from "components/employee/EmployeeSummary";
 import EmployeeHeader from "components/employee/Header/EmployeeHeader";
-import EmployeeSidePane from "components/employee/SidePane/EmployeeSidePane";
 import EmployeeTable from "components/employee/Table/EmployeeTable";
-import PayrollFormModal from "components/payroll/Form/PayrollFormModal";
-import ConfirmDialog from "components/shared/Dialog/ConfirmDialog/ConfirmDialog";
-import { EmployeeFormPayload } from "hooks/employee/useEmployeeForm";
-import { PayrollFormPayload } from "hooks/payroll/usePayrollForm";
-import { translate } from "i18n/i18n";
 import { observer } from "mobx-react-lite";
 import { useStore } from "stores/StoreContext";
+import { getEmployeesSummary } from "utils/employeeStats";
 
 import { Box } from "@mui/material";
 
 const EmployeePage: React.FC = observer(() => {
-	const { employeeStore, payrollStore } = useStore();
+	const { employeeStore } = useStore();
+	const navigate = useNavigate();
 
 	useEffect(() => {
 		employeeStore.getAll();
 	}, [employeeStore]);
 
-	const handleFormSave = (payload: EmployeeFormPayload) =>
-		employeeStore.selectedEmployee
-			? employeeStore.update({ id: employeeStore.selectedEmployee.id, ...payload })
-			: employeeStore.create({ ...payload });
-
-	const handlePayrollSave = async (payload: PayrollFormPayload) => {
-		await payrollStore.create(payload);
-		employeeStore.closeDialog();
-	};
-
-	const handleDeleteConfirmed = () => {
-		if (employeeStore.selectedEmployee) {
-			employeeStore.delete(employeeStore.selectedEmployee.id);
-		}
-	};
-
-	const employeesCount = useMemo(() => {
-		if (employeeStore.filteredEmployees === "loading") {
-			return "";
-		}
-
-		return employeeStore.filteredEmployees.length.toString();
-	}, [employeeStore.filteredEmployees]);
-
-	const { dialogMode } = employeeStore;
-	const dialogKind = dialogMode.kind;
+	const employees = employeeStore.filteredEmployees;
+	const summary =
+		employeeStore.allEmployees === "loading"
+			? null
+			: getEmployeesSummary(employeeStore.allEmployees);
 
 	return (
 		<Box>
 			<EmployeeHeader
 				searchValue={employeeStore.searchTerm}
 				selectedStatus={employeeStore.filterStatus}
-				titleCount={employeesCount}
 				onSearch={(value) => employeeStore.setSearch(value)}
 				onStatusChange={(value) => employeeStore.setFilterStatus(value)}
 				onCreate={employeeStore.openCreate}
 			/>
 
+			{summary && <EmployeeSummary summary={summary} />}
+
 			<EmployeeTable
-				data={employeeStore.filteredEmployees}
+				data={employees}
 				pagination
 				onSort={employeeStore.setSort}
 				onEdit={employeeStore.openEdit}
 				onDelete={employeeStore.openDelete}
 				onPayment={employeeStore.openPayment}
-				onViewDetails={employeeStore.openDetails}
+				onViewDetails={(employee) => navigate(`/employees/${employee.id}`)}
 			/>
 
-			<EmployeeFormModal
-				isOpen={dialogKind === "form"}
-				isSaving={employeeStore.isSaving}
-				employee={employeeStore.selectedEmployee}
-				onClose={employeeStore.closeDialog}
-				onSave={handleFormSave}
-			/>
-
-			<PayrollFormModal
-				isOpen={dialogKind === "payment"}
-				isSaving={payrollStore.isSaving}
-				mode={employeeStore.selectedEmployee}
-				onClose={employeeStore.closeDialog}
-				onSave={handlePayrollSave}
-			/>
-
-			<EmployeeSidePane
-				open={dialogKind === "details"}
-				employee={employeeStore.selectedEmployee}
-				onClose={employeeStore.closeDialog}
-				onEdit={employeeStore.openEdit}
-				onDelete={employeeStore.openDelete}
-				onPayment={employeeStore.openPayment}
-			/>
-
-			<ConfirmDialog
-				isOpen={dialogKind === "delete"}
-				title={translate("common.deleteTitle")}
-				content={translate("employee.deleteConfirmation", {
-					employeeName: employeeStore.selectedEmployee?.name ?? "",
-				})}
-				onConfirm={handleDeleteConfirmed}
-				onCancel={employeeStore.closeDialog}
-			/>
+			<EmployeeDialogs />
 		</Box>
 	);
 });
